fix: refactor date module
