+++ src/1shared/lib/date/model/date.const.ts
@@ -1,0 +1,18 @@
+import { TMonth, TTimePeriod } from "./date.types";
+
+export const MONTHS: TMonth[] = [
+  "jan",
+  "feb",
+  "mar",
+  "apr",
+  "may",
+  "june",
+  "july",
+  "aug",
+  "sept",
+  "oct",
+  "nov",
+  "dec",
+] as const;
+
+export const TIME_PERIODS: TTimePeriod[] = ["year", ...MONTHS] as const;
