+++ src/1shared/lib/date/model/date.types.ts
@@ -1,0 +1,5 @@
+export type TMonth = "jan" | "feb" | "mar" | "apr" | "may" | "june" | "july" | "aug" | "sept" | "oct" | "nov" | "dec";
+
+export type TTimePeriod = "year" | TMonth;
+
+export type TQuartalNumber = 1 | 2 | 3 | 4;
