+++ src/4widgets/monthPlan/ui/MonthPlanTable.tsx
@@ -1,8 +1,7 @@
 "use client";
 import { FC, Fragment } from "react";
 
-import { TTimePeriod } from "@/1shared/lib/date";
-import { translateRuTimePeriod } from "@/1shared/lib/date/locale";
+import { getTimePeriodFromString, TTimePeriod, translateRuTimePeriod } from "@/1shared/lib/date";
 import {
   getFactNormTimeFieldByTimePeriod,
   getFactTimeFieldByTimePeriod,
@@ -166,7 +165,8 @@ export const MonthPlanTable: FC<IMonthPlanTableProps> = ({
                 })}
                 <td>
                   {pprData[planWorkField].undoneTransfers?.map((transfer, index) => {
-                    const month = translateRuTimePeriod(transfer.fieldTo);
+                    const timePeriod = getTimePeriodFromString(transfer.fieldTo);
+                    const month = timePeriod && translateRuTimePeriod(timePeriod);
                     const text = transfer.value >= 0 ? `${transfer.value} на ${month}` : `${transfer.value} с ${month}`;
                     return <div key={index}>{text}</div>;
                   })}
